⚡️ :: update worker factory
